Change blog posts permalinks
Permalinks are now prefixed by `/blog`.
The publication date is also omitted.

Including the date in the URL makes sense when covering news.
But if posts are meant to stay current and be long-lived then it's
logical to omit the publication date.

In any case, the date of each post will be included in the post body.