Strip
Fix mind-boggling error that did not occur on Windows but occurred in a
GitHub Actions container (Ubuntu).

On `bundle exec jekyll build` the build failed with:

Liquid Exception: No such file or directory @ rb_sysopen - ./assets/blocks/animating-bars-going-down.html in /home/runner/work/guilhermesimoes.github.io/guilhermesimoes.github.io/_layouts/block.html

/home/runner/work/guilhermesimoes.github.io/guilhermesimoes.github.io/_plugins/root_include.rb:22:in `read': No such file or directory @ rb_sysopen - ./assets/blocks/animating-bars-going-down.html  (Errno::ENOENT)

The thing is, the file did exist. It was a rogue whitespace at the end
of the file path that was causing the error. But not on Windows.